requirements.txt
fastapi
uvicorn[standard]
gunicorn
python-dotenv
pydantic
pydantic-settings
langchain
langchain-openai
langchain-community
langgraph
langsmith
psycopg[binary]
aiosmtplib
aioimaplib
requests
duckduckgo-search
tavily-python
langchain-tavily
beautifulsoup4
email-validator
pytest
httpx
ddgs
trafilatura
playwright